+++ src/gfxDrawColors.cpp
@@ -19,11 +19,16 @@ namespace gfxDraw {
 
 // ===== ARGB class members =====
 
+
+// RGB+A Color
 ARGB::ARGB(uint8_t r, uint8_t g, uint8_t b, uint8_t a)
   : Red(r), Green(g), Blue(b), Alpha(a){};
 
+// raw 32 bit color with no transparency
 ARGB::ARGB(uint32_t color)
-  : raw(color){};
+  : raw(color) {
+  Alpha = 0xFF;
+};
 
 constexpr bool ARGB::operator==(const ARGB &col2) {
   return (raw == col2.raw);
@@ -36,7 +41,7 @@ constexpr bool ARGB::operator!=(const ARGB &col2) {
 /// @brief Convert into a 3*8 bit value using #RRGGBB.
 /// @return color value.
 uint32_t ARGB::toColor24() {
-  return (raw & 0x0FFF);
+  return (raw & 0x0FFFFFF);
 }
 
 // Convert into a 16 bit value using 5(R)+6(G)+5(B)